[modify] create auth entity with user entity
