+++ app/persistence/model/auth_model.py
@@ -40,4 +40,5 @@ class AuthModel(db.Model):
             expired_at=self.expired_at,
             created_at=self.created_at,
             updated_at=self.updated_at,
+            user=self.user.to_entity() if self.user else None,
         )
